Added a lower limit to the size of any link added to the list
Edge Case 2 results in an empty link being placed within the list. I decided to remove this from happening as generally, people are not expecting an empty string as a link. I chose a lower bound of 3 to allow links that have 1 letter, and then 2 extra letters after the period.