Added contain period condition to adding a string.
In edge case 3, there is a brackets and parentheses, but no link. This resulted in a non-link to be placed within the list.

As a band-aid solution, I added a new condition requiring the string to have a period in it to be added to the list, as I believe all links have a period in it. However, this doesn't really account for nonlink statements with periods in it, but still in a parentheses.